--- a/.github/CODEOWNERS
+++ b/.github/CODEOWNERS
@@ -4,4 +4,4 @@
 *.md    @fadeev @ilgooz @barriebyron @ignite-hq/growth-ecodev
 
 # Primary repo maintainers
-*       @fadeev @ilgooz @lubtd @Pantani @ivanovpetr
+*       @fadeev @ilgooz @lubtd @Pantani @ivanovpetr @jeronimoalbi