--- a/ignite/cmd/chain_build.go
+++ b/ignite/cmd/chain_build.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"path/filepath"
 
-	"github.com/ignite-hq/cli/ignite/chainconfig"
-	"github.com/ignite-hq/cli/ignite/pkg/cache"
 	"github.com/ignite-hq/cli/ignite/pkg/chaincmd"
 	"github.com/ignite-hq/cli/ignite/pkg/cliui/colors"
 	"github.com/ignite-hq/cli/ignite/services/chain"
@@ -74,11 +72,7 @@ func chainBuildHandler(cmd *cobra.Command, _ []string) error {
 		return err
 	}
 
-	cacheRootDir, err := chainconfig.ConfigDirPath()
-	if err != nil {
-		return err
-	}
-	cacheStorage, err := cache.NewStorage(cacheRootDir)
+	cacheStorage, err := newCache()
 	if err != nil {
 		return err
 	}
@@ -112,5 +106,5 @@ func chainBuildHandler(cmd *cobra.Command, _ []string) error {
 		fmt.Printf("🗃  Binary built at the path: %s\n", colors.Info(binaryPath))
 	}
 
-	return cacheStorage.Close()
+	return nil
 }