--- a/ignite/cmd/chain_build.go
+++ b/ignite/cmd/chain_build.go
@@ -72,15 +72,9 @@ func chainBuildHandler(cmd *cobra.Command, _ []string) error {
 		return err
 	}
 
-	cacheStorage, err := newCache()
+	cacheStorage, err := newCache(cmd)
 	if err != nil {
 		return err
-	}
-
-	if flagGetClearCache(cmd) {
-		if err := cacheStorage.Clear(); err != nil {
-			return err
-		}
 	}
 
 	if isRelease {