--- a/ignite/cmd/cmd.go
+++ b/ignite/cmd/cmd.go
@@ -9,12 +9,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/ignite-hq/cli/ignite/pkg/cliui"
-
 	"github.com/fatih/color"
-	"github.com/spf13/cobra"
-	flag "github.com/spf13/pflag"
-
+	"github.com/ignite-hq/cli/ignite/chainconfig"
+	"github.com/ignite-hq/cli/ignite/pkg/cache"
+	"github.com/ignite-hq/cli/ignite/pkg/cliui"
 	"github.com/ignite-hq/cli/ignite/pkg/cosmosaccount"
 	"github.com/ignite-hq/cli/ignite/pkg/cosmosver"
 	"github.com/ignite-hq/cli/ignite/pkg/gitpod"
@@ -23,6 +21,8 @@ import (
 	"github.com/ignite-hq/cli/ignite/services/chain"
 	"github.com/ignite-hq/cli/ignite/services/scaffolder"
 	"github.com/ignite-hq/cli/ignite/version"
+	"github.com/spf13/cobra"
+	flag "github.com/spf13/pflag"
 )
 
 const (
@@ -274,4 +274,12 @@ https://docs.ignite.com/migration`, sc.Version.String(),
 
 func printSection(session cliui.Session, title string) error {
 	return session.Printf("------\n%s\n------\n\n", title)
+}
+
+func newCache() (cache.Storage, error) {
+	cacheRootDir, err := chainconfig.ConfigDirPath()
+	if err != nil {
+		return cache.Storage{}, err
+	}
+	return cache.NewStorage(cacheRootDir)
 }