--- a/ignite/cmd/generate_dart.go
+++ b/ignite/cmd/generate_dart.go
@@ -27,6 +27,12 @@ func generateDartHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
+	if flagGetClearCache(cmd) {
+		if err := c.CacheStorage.Clear(); err != nil {
+			return err
+		}
+	}
+
 	if err := c.Generate(cmd.Context(), chain.GenerateDart()); err != nil {
 		return err
 	}