--- a/ignite/cmd/generate_dart.go
+++ b/ignite/cmd/generate_dart.go
@@ -3,8 +3,6 @@ package ignitecmd
 import (
 	"fmt"
 
-	"github.com/ignite-hq/cli/ignite/chainconfig"
-	"github.com/ignite-hq/cli/ignite/pkg/cache"
 	"github.com/ignite-hq/cli/ignite/pkg/cliui/clispinner"
 	"github.com/ignite-hq/cli/ignite/services/chain"
 	"github.com/spf13/cobra"
@@ -28,11 +26,7 @@ func generateDartHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	cacheRootDir, err := chainconfig.ConfigDirPath()
-	if err != nil {
-		return err
-	}
-	cacheStorage, err := cache.NewStorage(cacheRootDir)
+	cacheStorage, err := newCache()
 	if err != nil {
 		return err
 	}
@@ -50,5 +44,5 @@ func generateDartHandler(cmd *cobra.Command, args []string) error {
 	s.Stop()
 	fmt.Println("⛏️  Generated Dart client.")
 
-	return cacheStorage.Close()
+	return nil
 }