--- a/ignite/cmd/generate_dart.go
+++ b/ignite/cmd/generate_dart.go
@@ -26,15 +26,9 @@ func generateDartHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	cacheStorage, err := newCache()
+	cacheStorage, err := newCache(cmd)
 	if err != nil {
 		return err
-	}
-
-	if flagGetClearCache(cmd) {
-		if err := cacheStorage.Clear(); err != nil {
-			return err
-		}
 	}
 
 	if err := c.Generate(cmd.Context(), cacheStorage, chain.GenerateDart()); err != nil {