--- a/ignite/cmd/generate_go.go
+++ b/ignite/cmd/generate_go.go
@@ -26,6 +26,12 @@ func generateGoHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
+	if flagGetClearCache(cmd) {
+		if err := c.CacheStorage.Clear(); err != nil {
+			return err
+		}
+	}
+
 	if err := c.Generate(cmd.Context(), chain.GenerateGo()); err != nil {
 		return err
 	}