--- a/ignite/cmd/generate_go.go
+++ b/ignite/cmd/generate_go.go
@@ -3,8 +3,6 @@ package ignitecmd
 import (
 	"fmt"
 
-	"github.com/ignite-hq/cli/ignite/chainconfig"
-	"github.com/ignite-hq/cli/ignite/pkg/cache"
 	"github.com/ignite-hq/cli/ignite/pkg/cliui/clispinner"
 	"github.com/ignite-hq/cli/ignite/services/chain"
 	"github.com/spf13/cobra"
@@ -27,11 +25,7 @@ func generateGoHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	cacheRootDir, err := chainconfig.ConfigDirPath()
-	if err != nil {
-		return err
-	}
-	cacheStorage, err := cache.NewStorage(cacheRootDir)
+	cacheStorage, err := newCache()
 	if err != nil {
 		return err
 	}
@@ -49,5 +43,5 @@ func generateGoHandler(cmd *cobra.Command, args []string) error {
 	s.Stop()
 	fmt.Println("⛏️  Generated go code.")
 
-	return cacheStorage.Close()
+	return nil
 }