--- a/ignite/cmd/generate_go.go
+++ b/ignite/cmd/generate_go.go
@@ -25,15 +25,9 @@ func generateGoHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	cacheStorage, err := newCache()
+	cacheStorage, err := newCache(cmd)
 	if err != nil {
 		return err
-	}
-
-	if flagGetClearCache(cmd) {
-		if err := cacheStorage.Clear(); err != nil {
-			return err
-		}
 	}
 
 	if err := c.Generate(cmd.Context(), cacheStorage, chain.GenerateGo()); err != nil {