--- a/ignite/cmd/generate_vuex.go
+++ b/ignite/cmd/generate_vuex.go
@@ -28,6 +28,12 @@ func generateVuexHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
+	if flagGetClearCache(cmd) {
+		if err := c.CacheStorage.Clear(); err != nil {
+			return err
+		}
+	}
+
 	if err := c.Generate(cmd.Context(), chain.GenerateVuex()); err != nil {
 		return err
 	}