--- a/ignite/cmd/generate_vuex.go
+++ b/ignite/cmd/generate_vuex.go
@@ -3,10 +3,11 @@ package ignitecmd
 import (
 	"fmt"
 
-	"github.com/spf13/cobra"
-
+	"github.com/ignite-hq/cli/ignite/chainconfig"
+	"github.com/ignite-hq/cli/ignite/pkg/cache"
 	"github.com/ignite-hq/cli/ignite/pkg/cliui/clispinner"
 	"github.com/ignite-hq/cli/ignite/services/chain"
+	"github.com/spf13/cobra"
 )
 
 func NewGenerateVuex() *cobra.Command {
@@ -28,18 +29,27 @@ func generateVuexHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
+	cacheRootDir, err := chainconfig.ConfigDirPath()
+	if err != nil {
+		return err
+	}
+	cacheStorage, err := cache.NewStorage(cacheRootDir)
+	if err != nil {
+		return err
+	}
+
 	if flagGetClearCache(cmd) {
-		if err := c.CacheStorage.Clear(); err != nil {
+		if err := cacheStorage.Clear(); err != nil {
 			return err
 		}
 	}
 
-	if err := c.Generate(cmd.Context(), chain.GenerateVuex()); err != nil {
+	if err := c.Generate(cmd.Context(), cacheStorage, chain.GenerateVuex()); err != nil {
 		return err
 	}
 
 	s.Stop()
 	fmt.Println("⛏️  Generated vuex stores.")
 
-	return nil
+	return cacheStorage.Close()
 }