--- a/ignite/cmd/generate_vuex.go
+++ b/ignite/cmd/generate_vuex.go
@@ -27,15 +27,9 @@ func generateVuexHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	cacheStorage, err := newCache()
+	cacheStorage, err := newCache(cmd)
 	if err != nil {
 		return err
-	}
-
-	if flagGetClearCache(cmd) {
-		if err := cacheStorage.Clear(); err != nil {
-			return err
-		}
 	}
 
 	if err := c.Generate(cmd.Context(), cacheStorage, chain.GenerateVuex()); err != nil {