--- a/ignite/cmd/network_chain_install.go
+++ b/ignite/cmd/network_chain_install.go
@@ -21,6 +21,8 @@ func NewNetworkChainInstall() *cobra.Command {
 		Args:  cobra.ExactArgs(1),
 		RunE:  networkChainInstallHandler,
 	}
+
+	flagSetClearCache(c)
 	c.Flags().AddFlagSet(flagNetworkFrom())
 	return c
 }
@@ -28,6 +30,17 @@ func NewNetworkChainInstall() *cobra.Command {
 func networkChainInstallHandler(cmd *cobra.Command, args []string) error {
 	session := cliui.New()
 	defer session.Cleanup()
+
+	cacheStorage, err := newCache()
+	if err != nil {
+		return err
+	}
+
+	if flagGetClearCache(cmd) {
+		if err := cacheStorage.Clear(); err != nil {
+			return err
+		}
+	}
 
 	nb, err := newNetworkBuilder(cmd, CollectEvents(session.EventBus()))
 	if err != nil {
@@ -55,7 +68,7 @@ func networkChainInstallHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	binaryName, err := c.Build(cmd.Context())
+	binaryName, err := c.Build(cmd.Context(), cacheStorage)
 	if err != nil {
 		return err
 	}