--- a/ignite/cmd/network_chain_install.go
+++ b/ignite/cmd/network_chain_install.go
@@ -31,15 +31,9 @@ func networkChainInstallHandler(cmd *cobra.Command, args []string) error {
 	session := cliui.New()
 	defer session.Cleanup()
 
-	cacheStorage, err := newCache()
+	cacheStorage, err := newCache(cmd)
 	if err != nil {
 		return err
-	}
-
-	if flagGetClearCache(cmd) {
-		if err := cacheStorage.Clear(); err != nil {
-			return err
-		}
 	}
 
 	nb, err := newNetworkBuilder(cmd, CollectEvents(session.EventBus()))