--- a/ignite/cmd/network_chain_prepare.go
+++ b/ignite/cmd/network_chain_prepare.go
@@ -27,6 +27,7 @@ func NewNetworkChainPrepare() *cobra.Command {
 		RunE:  networkChainPrepareHandler,
 	}
 
+	flagSetClearCache(c)
 	c.Flags().BoolP(flagForce, "f", false, "Force the prepare command to run even if the chain is not launched")
 	c.Flags().AddFlagSet(flagNetworkFrom())
 	c.Flags().AddFlagSet(flagSetKeyringBackend())
@@ -40,6 +41,17 @@ func networkChainPrepareHandler(cmd *cobra.Command, args []string) error {
 	defer session.Cleanup()
 
 	force, _ := cmd.Flags().GetBool(flagForce)
+
+	cacheStorage, err := newCache()
+	if err != nil {
+		return err
+	}
+
+	if flagGetClearCache(cmd) {
+		if err := cacheStorage.Clear(); err != nil {
+			return err
+		}
+	}
 
 	nb, err := newNetworkBuilder(cmd, CollectEvents(session.EventBus()))
 	if err != nil {
@@ -78,7 +90,7 @@ func networkChainPrepareHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	if err := c.Prepare(cmd.Context(), genesisInformation); err != nil {
+	if err := c.Prepare(cmd.Context(), cacheStorage, genesisInformation); err != nil {
 		return err
 	}