# Changes made
This PR adds a new key-value cache store to the CLI which can be used to cache values so that building and generation can be sped up significantly.

After analyzing the bottlenecks I realized that the main places where a lot of time was being spent was on analyzing code and dependencies. The changes are therefore focused on caching those values and making sure they get invalidated whenever changes occur.

I also added a --clear-cache flag to the relevant cli commands as a simple way of making sure caching isn't causing problems anywhere. I haven't been able to create any such problems during testing however.

The performance improvements work pretty much across the board for build, serve and even some scaffold commands to some extent. 

Closes #2383

# Measurments

## First serve after scaffold

Baseline with "old" cli:
```bash
$ time ./ignite-before c serve
Cosmos SDK's version is: stargate - v0.45.3

🛠️  Building proto...
etc...

real	0m42,947s
user	2m18,881s
sys	0m17,040s
```

With cache:
```bash
$ time ignite c serve
Cosmos SDK's version is: stargate - v0.45.3

🛠️  Building proto...
etc...
real	0m17,236s
user	1m1,851s
sys	0m7,525s
```

## Build

Baseline 1 with `go build` in cmd/chaind
```bash
$ time go build
real	0m2,459s
user	0m4,901s
sys	0m0,619s
```

`chain build` baseline with "old" cli:
```bash
$ time ./ignite-before c build
Cosmos SDK's version is: stargate - v0.45.3

🛠️  Building proto...
etc...
real	0m33,573s
user	2m10,636s
sys	0m11,209s
```

`chain build` with cache and no changes
```bash
$ time ignite c build
Cosmos SDK's version is: stargate - v0.45.3

🛠️  Building proto...
etc...
real	0m2,864s
user	0m6,694s
sys	0m1,175s
```

`chain build` with changes in chain module (no proto changes)
```bash
$ time ignite c build
Cosmos SDK's version is: stargate - v0.45.3

🛠️  Building proto...
etc...
real	0m5,743s
user	0m10,112s
sys	0m1,520s
```

`chain build` with changes in proto only
```bash
$ time ignite c build
Cosmos SDK's version is: stargate - v0.45.3

🛠️  Building proto...
etc...
real	0m8,614s
user	0m14,560s
sys	0m1,730s
```

## Generate with --proto-all-modules

Without cache
```bash
$ time ignite g vuex --proto-all-modules --clear-cache -y
⛏️  Generated vuex stores.

real	0m38,508s
user	3m30,369s
sys	0m13,932s
```

With cache
```bash
$ time ignite g vuex --proto-all-modules -y
⛏️  Generated vuex stores.

real	0m2,259s
user	0m3,923s
sys	0m0,613s
```