@bjaanes 2022-05-05
@ilgooz PR ready for review. I've tried to explain most of my changes, but let me know if you have any questions/concerns and if you think the performance improvements are sufficient.