@bjaanes 2022-05-05
Added benefit I just realized:

![image](https://user-images.githubusercontent.com/1344105/167002438-08336bc8-efae-4088-8729-00b23205fad0.png)

![image](https://user-images.githubusercontent.com/1344105/167001977-1a433d2c-0205-4e1e-974b-238a3258ab66.png)