@ilgooz 2022-05-08
We might consider deprecating `--proto-all-modules`. What do you think?