@bjaanes 2022-05-08
> We might consider deprecating `--proto-all-modules`. What do you think?

With my testing it seems like it could be doable. After it was generated things went fast anyway, so could simply things quite a bit.