@bjaanes 2022-05-17
@ilgooz In reference to the solution you outlined to avoid the Close problem. I've implemented it now and it seems to work very well. It makes the package a lot easier to use and doesn't seem to add any issues in terms of performance or otherwise.