@bjaanes 2022-05-17
> Just noticed one issue:
> 
> * serve a chain
> * `touch app/newfile.go` > this won't trig a binary rebuild

Hmm, not sure why. But as soon as you change it, it triggers.